# Hyperloglog ARM NEON SIMD optimization