@xbasel 2025-03-21
> 10 times faster is a pretty good improvement. :)
> 
> I didn't read the NEON code carefully because I'm not familiar with it. Is the logic basically the same as the one for AVX2?

It is similar. NEON vectors are 128 bit, AVX2 is 256 bit. The padding and lookup is a bit different in AVX2.
The execution time  of pfmerge ~**14.7x faster**. The end to end is ~10x faster.