@xbasel 2025-04-30
> LGTM.
> 
> I didn't review the actual SIMD code very carefully, but we have a test to compare the results with/without SIMD so I think it's safe.
> 
> I'll wait for @lipzhu's approval before merge.

The code is actually already being tested with  SIMD vs Scalar in `tests/unit/hyperloglog.tcl`. See `PFMERGE results with simd` test.
All you need to do is to run the test on ARM.

```
➜  valkey git:(hll_neon) ✗ ./runtest --single unit/hyperloglog --only "PFMERGE results with simd"
Cleanup: may take some time... OK
..
[ok]: PFMERGE results with simd (457 ms)
..
[1/1 done]: unit/hyperloglog (1 seconds)

                   The End

Execution time of different units:
  1 seconds - unit/hyperloglog

\o/ All tests passed without errors!

Cleanup: may take some time... OK
➜  valkey git:(hll_neon) ✗ cat /proc/cpuinfo
processor	: 0
BogoMIPS	: 2100.00
Features	: fp asimd evtstrm aes pmull sha1 sha2 crc32 atomics fphp asimdhp cpuid asimdrdm jscvt fcma lrcpc dcpop sha3 sm3 sm4 asimddp sha512 sve asimdfhm dit uscat ilrcpc flagm ssbs paca pacg dcpodp svei8mm svebf16 i8mm bf16 dgh rng
CPU implementer	: 0x41
CPU architecture: 8
```

This is an ARMv8-A 64-bit CPU (and supports NEON).