@zuiderkwast 2025-04-30
> The code is actually already being tested with SIMD vs Scalar in `tests/unit/hyperloglog.tcl`. See `PFMERGE results with simd` test.
> All you need to do is to run the test on ARM.

Yeah, I know. That's why I said "we have a test to compare the results with/without SIMD so I think it's safe".